requirements.txt
requests
openpyxl
pandas
pytz
colorama
Pillow
pytesseract